Add files via upload
Robo Speaker 2.0
A simple Python-based text-to-speech application using the pyttsx3 library. This project allows users to input text, which the program then speaks out loud. The application continues to listen for input until the user types "stop," at which point it bids farewell and exits. Perfect for exploring basic text-to-speech functionality in Python.